+++ tvlm/analysis/cfgNode.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <set>
-#include "../il.h"
+#include "tvlm/tvlm/il/il.h"
 
 using Instruction = ::tvlm::Instruction;
 using Function  = ::tvlm::Function;
