+++ tvlm/tvlm.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "il_builder.h"
-#include "il.h"
+#include "tvlm/tvlm/il/il_builder.h"
+#include "tvlm/tvlm/il/il.h"
 
 namespace tvlm {
